Remove SyncScanInterpreter
