+++ rust/nasl-interpreter/src/lib.rs
@@ -28,9 +28,9 @@ pub use interpreter::Interpreter;
 pub use scanner::ExecuteError;
 
 pub use scanner::DefaultScannerStack;
+pub use scanner::ScanRunner;
 pub use scanner::Scanner;
 pub use scanner::ScannerStackWithStorage;
-pub use scanner::SyncScanInterpreter;
 // we expose the other libraries to allow users to use them without having to import them
 pub use nasl_builtin_std::{nasl_std_functions, ContextFactory, RegisterBuilder};
 pub use nasl_builtin_utils::{
